Use picker style for Require Authentication (#506)
